// src/app/services/data/model/notifications.ts
export class LunchAccountBalance {
  value: any;
  option: string;
}


export class Notifications {
  AccountBalance: any;
  LowMealBalance: boolean;
  DayofMonth: any;
  BalanceReminder: boolean;
  FundAccountBalance: any;
  LowFundBalance: boolean;
  FundDayofMonth: any;
  FundBalanceReminder: boolean;
  DaysNoticeDue: any;
  FeeDue: boolean;
  DaysNoticePayment: any;
  UpcomingPayment: boolean;
  DaysOverdueFee: any;
  OverdueFee: boolean;
  DayofMonthStatement: any;
  MonthlyStatement: boolean;
}